requirements.txt
ics==0.7.2
python_dateutil==2.9.0
python_dateutil==2.9.0.post0
Requests==2.32.3
